[datastore] add save_users_docs method
from orchestrator, we need to periodically save user_docs for all users
It's more efficient (cost and latency) to make vectorial requests instead
of one set of requests by user (mainly with remote_api).
(Given that we target one billion users for the end of 2016 :-) )
NB: not vectorial method is still used for dummy values and unit tests. It might be eventually deleted