[Orchestrator] add unit test fix bugs + refactor orchestrate()
To make unit test, we inject Scraper and TopicModeller in _orchestrate method (two modules with complex logic and using files). To do so we wrap scrap method by an injectable Scraper class. This gives also the possibility to inject a scraper class replaying documents saved on drive instead of real scraping

We inject also DocSaver (file access painfull to test), this allows also to call orchestrator with saver doing nothing if we want (or saving in database).

To test orchestrator a needed to init mock of datastore as for test_dal. This created duplicated code, I factorized this in a testhealpers module common package.

I also fix various bugs in orchestrator I introduced previously on adapting it to learner (and did not test back then)