[scraper] fix test on max length
-make to assert (not two long, too long) with identical construction, to be sure the one too long fails for this specific reason (thus making test more robust)

-I found that len(unicode), which is the method called by  return the length of the byte array, so a call to _is_size_reasonable
gives an higher weight on special char (2,3 depending on the char). We just want a basic check to remove aberrant so I let it as is (the alternative is to entiery decode string). But I need to take it into account in unit test because I explicitly choose de character 'é' so the string is detected as utf8 instead of ascii