Adapt dal clients to use a dal instance
Dal now keep a connection to datastore gcloud client and is a class
So clients of dal needs to instanciate a Dal object.

This is not strictly necessary now because "client" is now a static class field and not an instance field of Dal class
But I put it as a class field afterwards for performance. I kept instances of Dal everywhere anyway because it's more flexible
(to make a specific client by instance for example)