[functests] Retry actions on browser 2 times if timeout
We get random timeouts in Travis on actions on selenium/phantomJS
This seems to be a problem specific to Docker/phantomJS mix. More details in the issue below.

Retry is the approch of the Pull below. We apply it to get pages and click links (where I
spotted the issues)

I don't know why we did not get those issues before Flask migration...

problem timeout in tavis
https://github.com/travis-ci/travis-ci/issues/3251
https://github.com/spacetelescope/asv/pull/290