[python3] move main.py and gae app root directory
implicit relative import as been removed in python3
We cannot start a script inside a package and make it
import modules in the same package.

The solution is to follow structure as here:
https://github.com/pallets/flask/tree/master/examples/blueprintexample
the main.py is ouside the tested package server.
So it's the wall gae app that is move one directory above
That's imply to skip tests explicitly in app.yaml to not
distribute them