Move main.py and gae app root directory
implicit relative import as been removed in python3 (was already a pylint warning)
We cannot start a script inside a package and make it
import modules in the same package.

The solution is to follow structure as here:
https://github.com/pallets/flask/tree/master/examples/blueprintexample
the main.py is ouside the tested package server.
So it's the wall gae app that is move one directory above
This imply to skip tests explicitly in app.yaml to not
distribute them

We also upgrade dependency versions in requirements for appengine server