[TopicModeller] manage unclassifiable doc + optimize
1) classify method return two elt tuple (is_ok, vector). I detected some docs
that are not classified (all vector values at zero), it leads to NaN grade when computing cosinus
returning a status enable scrap_and_learn to deal with it

2) topics field init is very slow because it calls nb_topics time "self._lda.show_topic()" which is slow.
-need for this field has been removed in classify function
-field is now lazy loaded (its only use is to init feature_set with relevant topic names)
Those two devs allows to not use this field in coming functional test of scrap_and_learn

For the record: self._lda.show_topic() is slow because it calls a function "Lambda()" internally for each call.
This would better solved inside gensim by a vectorized function on topics because lambda() doesn't depends
on "topicid" parameter.