Add production "script function" for scrap_and_learn + functest
To run scrap_learn in a container, we need to make it a script (instead of a function with parameters), but in the same time we need to test it (without container yet)

Solution is to make a parameterless function that can be tested, and that takes its input from command line.
then call the function in a one line script (launch_scrap_and_learn) and test it in functest

To make test possible:
-scrap_learn can now manage a given number of doc
-it manages unclassifiable docs
-"script" function can manage vcr cassette instead of doing internet requests

Contrary to all unit tests, the functest uses production (versioned) topicModeller. It should be calling "containerized" scipt in the future