don't mock localhost requests with vcrpy
localhost requests are used for local datastore.

It worked without it because if first request of dal is
performed outside cassette context (which was the case), the nexts are too.

(but the fix is mandatory for scrap_learn script in docker image where first
dal request is already in cassette context)