Ensure user is unique by email while keeping it modifiable
Email should be unique as it is used as idendifier of a user
This was the reason why it was the id of user entity
At the same time it can be modified, while entity id can't
The solution is to make another table, whose id is email and value is key to
user entity, so related entities can point to unmodifiable user key.

It means was is guaranteed by database is unicity but an existing user could be overriden
if existance is not checked before (or because of race condition). This check is done on client side
(at user creation in handlers) because dal cannot offer guaranty

It also means that some user can become "zombi" is the sense that no email points to it
get_all_user has been modified to filter them (needed when we launch a unit tests several times)
get_all_user modificaton has not been unit tested as all this will change with userAnonym
get_user_actions_on_docs has also been adapted as it relied on the fact that user id was its email
get_all_user test has been modified to not rely on email field

unicity technique is described here:
https://www.reddit.com/r/AppEngine/comments/41fr30/storing_unique_users_by_email_in_datastore/+&cd=5&hl=fr&ct=clnk&gl=fr